requirements.txt
fastapi
uvicorn
pydantic
google-adk
dotenv